Prepare geoRouter file structure
