+++ map/src/class/geoRouter/flushState.js
@@ -1,0 +1,7 @@
+export function flushState() {
+    if (this.setter) {
+        this.setter((previous) => Object.assign({}, previous, this)); // queue
+    } else {
+        console.error('geoRouter incorrect flushState() call');
+    }
+}
