+++ map/src/class/geoRouter/setPause.js
@@ -1,0 +1,8 @@
+
+export function setPause({ setter = null, pause }) {
+    if (pause === true || pause === false) {
+        this.initSetter(setter);
+        this.paused = pause;
+        this.flushState();
+    }
+}
