+++ map/src/context/RoutingManager.js
@@ -2,8 +2,6 @@ import TracksManager from "./TracksManager";
 import _ from "lodash";
 import TrackLayerProvider from "../map/TrackLayerProvider";
 import { apiGet } from '../util/HttpApi';
-import { copyObj } from "../util/Utils";
-import onlineRoutingProviders from "../generated/online-routing-providers.json";
 
 const STOP_CALC_ROUTING = 'stop';
 
@@ -75,31 +73,6 @@ function addSegmentToRouting(start, end, oldPoint, tempPolyline, segments) {
     }
     segments.push(segment);
     return segments;
-}
-
-function getColors() {
-    return {
-        'car': '#1976d2',
-        'truck': '#2F4F4F',
-        'motorcycle': '#f8931d',
-        'bicycle': '#9053bd',
-        'boat': '#08b5ff',
-        'horsebackriding': '#7f3431',
-        'pedestrian': '#d90139',
-        'ski': '#ffacdf',
-        'line': '#5F9EA0',
-        'moped': '#3e690e',
-        'train': '#a56b6f'
-    };
-}
-
-function addModes(data) {
-    data['line'] = { name: 'Line', params: {} };
-    return data;
-}
-
-function filterMode(data) {
-    return Object.fromEntries(Object.entries(data).filter(([key]) => !key.includes('rescuetrack')));
 }
 
 function osrmToFeaturesCollection(osrm) {
@@ -331,306 +304,6 @@ async function calculateGpxRoute({
     }
 }
 
-/*
-    OsmAnd provider structure and modifications: https://test.osmand.net/routing/routing-modes
-
-        1. OsmAnd type, router (key), url are pre-defined
-        2. OsmAnd provider is ONE only in providersOsmAnd[] array
-        3. OsmAnd profiles are converted from {} to profiles[] array
-        4. OsmAnd profile's .params are copied to .backup (used to reset params)
-
-    OSRM provider structure and modifications: https://test.osmand.net/online-routing-providers.json
-
-        1. OSRM type, name (as key), url, routes are loaded into providersOSRM[]
-        2. routes[] are renamed to profiles[], each profile .type is renamed to .key
-        3. finally, names are turned into pretty names by aliases (hardcoded)
-
-    Attributes usage:
-
-        1. providers[] usage: key as is, type, name (pretty), url (when profile has no url)
-        2. profiles[] usage: key as is, name (pretty), url/params are optional
-        3. current choosed: type, router(key), profile(key)
-
-    routeProviders comprises:
-
-        Currently selected provider: type, router (key), profile (key) -- will use these to keep in routing results;
-        Public getters() and ACTIONS() to work with;
-        Arrays of loaded (= available) providers;
-        Internal helper functions
-*/
-
-function initRouteProviders() {
-    const preloaded = {
-        type: 'osmand',
-        router: 'osmand',
-        profile: 'car',
-        name: 'OsmAnd Advanced',
-        url: `${process.env.REACT_APP_ROUTING_API_SITE}/routing/route`,
-    };
-
-    return {
-        setter: null,
-
-        // current
-        type: preloaded.type,
-        router: preloaded.router, // ref to providers.key
-        profile: preloaded.profile, // ref to profiles.key
-
-        // readiness
-        loaded: false,
-        paused: false,
-        isReady() { return this.loaded === true && this.paused === false },
-
-        // getters (use copy to prevent direct modify by parents)
-        allProviders() { return this._allProviders() }, // no-copy-need
-        getProvider(router = this.router) { return copyObj(this._getProvider(router)) },
-        allProfiles(router = this.router) { return copyObj(this._allProfiles(router)) },
-        getProfile(router = this.router, profile = this.profile) { return copyObj(this._getProfile(router, profile)) },
-        getParams(router = this.router, profile = this.profile) { return copyObj(this._getParams(router, profile)) },
-        getResetParams(router = this.router, profile = this.profile) { return copyObj(this._getResetParams(router, profile)) },
-        getURL(router = this.router, profile = this.profile) { return this._getURL(router, profile) },
-        getProviderByType(type = this.type) { return copyObj(this._getProviderByType(type)) },
-
-        // internals (strict and fast)
-        _allProviders() { return this.providersOSRM.concat(this.providersOsmAnd) }, // OSRM + OsmAnd
-        _getProvider(router) { return this._allProviders()?.find(r => r.key === router) ?? {} },
-        _allProfiles(router) { return this._getProvider(router)?.profiles ?? [] },
-        _getProfile(router, profile) { return this._allProfiles(router)?.find(p => p.key === profile) ?? {} },
-        _getParams(router, profile) { return this._getProfile(router, profile)?.params },
-        _getResetParams(router, profile) { return this._getProfile(router, profile)?.backup },
-        _getURL(router, profile) { return this._getProfile(router, profile)?.url ?? this._getProvider(router)?.url },
-        _getProviderByType(type) { return this._allProviders()?.find(r => r.type === type) ?? {} },
-
-        flushState() {
-            if (this.setter) {
-                this.setter((previous) => Object.assign({}, previous, this)); // queue
-            } else {
-                console.error('geoRouter incorrect flushState() call');
-            }
-        },
-
-        initSetter(setter = null) {
-            if (setter && setter !== this.setter) {
-                this.setter = setter;
-                this.flushState();
-                /*
-                 * there is 3 alternative ways:
-                 *
-                 * setter(o => Object.assign(o, { setter })); // next render
-                 * setter(o => { o.setter = setter; return {...o}; } ); // next render
-                 * this.setter = setter; setter(() => Object.assign({}, this)); // current render
-                 */
-            }
-            if (!this.setter) {
-                console.error('geoRouter incorrect initSetter() call');
-            }
-        },
-
-        // actions
-        setParams({ setter = null, opts }) {
-            /*
-                Saving profile params isn't easy process:
-
-                1. We can't modify _getProfile().params directly because allProviders() consist of joined arrays
-                2. We assume that opts are applicable to providersOsmAnd only (OSRM options aren't supported)
-                3. Therefore, we look into providersOsmAnd, rewrite params and call ctx-setter
-            */
-
-            this.initSetter(setter);
-
-            const router = this.router;
-            const profile = this.profile;
-
-            this.providersOsmAnd?.forEach(r => {
-                if (r.key === router) {
-                    r.profiles?.forEach(p => {
-                        if (p.key === profile) {
-                            p.params = opts;
-                        }
-                    });
-                }
-            });
-
-            this.flushState();
-        },
-
-
-        setPause({ setter = null, pause }) {
-            if (pause === true || pause === false) {
-                this.initSetter(setter);
-                this.paused = pause;
-                this.flushState();
-            }
-        },
-
-        choose({ setter = null, type = null, router = null, profile = null } = {}) {
-            /*
-                Switch to selected "router" and/or "profile":
-
-                1. use current router|profile when (any)===null
-                2. use first available when router|profile is not applicable
-                3. type is changed automatically according to applicated router
-
-                Examples:
-
-                choose({ [setter], profile }) - switch profile, keep router
-                choose({ [setter], router }) - switch router, try to keep profile
-                choose({ [setter], router, profile }) - switch both (if applicable)
-
-                Special case (used for searchParams / share route link):
-
-                choose({ [setter], type, [profile] }) - select 1st type's provider
-            */
-
-            this.initSetter(setter);
-
-            if (type && !router) {
-                router = this.getProviderByType(type)?.key ?? 'osmand';
-                // console.log('type', type, 'router', router);
-            }
-
-            // use current values if not defined
-            router = router ?? this.router;
-            profile = profile ?? this.profile;
-
-            const providers = this.allProviders();
-
-            if (!providers.find(r => r.key === router)) {
-                router = providers[0].key ?? 'osmand'; // fallback
-            }
-
-            const profiles = this.allProfiles(router);
-
-            if (!profiles.find(p => p.key === profile)) {
-                profile = profiles[0].key ?? 'car'; // fallback
-            }
-
-            if (router && profile) {
-                type = this.getProvider(router)?.type ?? 'osmand'; // fallback
-                // console.log('choosen:', type, router, profile);
-
-                this.type = type;
-                this.router = router;
-                this.profile = profile;
-
-                this.flushState();
-            }
-        },
-
-        // load and validate OSRM and OsmAnd routing providers
-        async loadProviders({ setter = null, creatingRouteMode = null, setCreatingRouteMode = null }) {
-            this.initSetter(setter);
-
-            let json = onlineRoutingProviders;
-
-            if(!json) {
-                const osrm = await apiGet(
-                    `${process.env.REACT_APP_USER_API_SITE}/online-routing-providers.json`,
-                    { apiCache: true }
-                );
-                if (osrm.ok) {
-                    json = osrm.data;
-                }
-            }
-
-            if (json && json?.providers && json?.providers[0]?.name) {
-
-                json.providers.forEach(p => { // routes to profiles
-                    p.name = p.webName ?? p.name;
-                    p.key = p.name; // name acts as key
-                    if (p.routes) {
-                        p.routes.forEach(r => { // type to key
-                            r.key = r.type;
-                            delete r.type;
-                        });
-                        p.profiles = p.routes;
-                        delete p.routes;
-                    }
-                });
-
-                this.providersOSRM = json.providers;
-                this.type = json.providers[0].type;
-                this.router = json.providers[0].key; // select first OSRM key and type
-                this.profile = json.providers[0]?.profiles[0]?.key; // select first profile
-            } else {
-                console.log('failed to load osrm providers');
-            }
-
-            // load OsmAnd provider as advanced solution
-            // TracksManager compatibility: provide OsmAnd to setCreatingRouteMode
-            // OsmAnd JSON profiles list is converted from Object to Array (for OSRM compatibility)
-            const osmand = await apiGet(
-                `${process.env.REACT_APP_ROUTING_API_SITE}/routing/routing-modes`,
-                { apiCache: true }
-            );
-            if (osmand.ok) {
-                try {
-                    const json = await osmand.json();
-
-                    // Tracks-routing compatibility
-                    if (json && setCreatingRouteMode) {
-                        let creatingData = _.cloneDeep(json);
-                        creatingData = filterMode(creatingData);
-                        creatingData = addModes(creatingData);
-                        setCreatingRouteMode( {
-                            mode: creatingRouteMode.mode,
-                            modes: creatingData,
-                            opts: creatingData[creatingRouteMode.mode]?.params,
-                            colors: getColors()
-                            }
-                        );
-                    }
-
-                    if (json) {
-                        // convert OsmAnd "profiles" {} to OSRM "profiles" [] array
-                        // note: sort, filter, additional profiles will be processed here
-                        // copy .params to .backup (used later to reset params in settings)
-                        const converted = [];
-                        Object.keys(json).forEach((k) => {
-                            if (json[k]?.params) {
-                                json[k].backup = copyObj(json[k]?.params);
-                            }
-                            converted.push(json[k]);
-                        });
-
-                        // update default OsmAnd provider with actual profiles
-                        this.providersOsmAnd = [
-                            Object.assign({},
-                                this.providersOsmAnd[0], // keep preloaded
-                                { profiles: converted } // add profiles
-                            )
-                        ];
-                    }
-                } catch { console.log('failed to load osmand profiles'); }
-
-            }
-
-            // set type/profile according to window.location.search
-            const searchParams = new URLSearchParams(window.location.search);
-            const type = searchParams.get('type');
-            const profile = searchParams.get('profile');
-            if (type && profile) {
-                this.choose({ type, profile });
-            }
-
-            this.loaded = true;
-
-            this.flushState();
-        },
-
-
-        /*
-            providersOSRM and providersOsmAnd are loaded by loadRouterProviders()
-            Nevertheless, providersOsmAnd might be used before they are loaded
-            Therefore, OsmAnd must be initialized with minimal-fallback data
-        */
-        providersOSRM: [],
-        providersOsmAnd: [Object.assign({}, preloaded,
-            { key: preloaded.router, profiles: [{ key: 'car', name: 'Car' }] })],
-
-    };
-}
-
 function routeModeCompatible(routeProviders) {
     const compatible = {
         mode: routeProviders.profile,
@@ -644,7 +317,6 @@ const RoutingManager = {
     getRoutingFromCash,
     validateRoutingCash,
     addSegmentToRouting,
-    initRouteProviders,
     routeModeCompatible,
     calculateGpxRoute,
     calculateRoute
